sf object of aerial imagery metadata for New Zealand.

This can be used to query aerial imagery coverage and metadata for New Zealand.

Usage

nz_aerial_imagery_metadata

Format

nz_aerial_imagery_metadata

A sf data frame with 230 rows and 8 columns:

layer_id

Layer ID of the aerial imagery

title

Title of the aerial imagery

abstract

Abstract of the aerial imagery

region

Region of the aerial imagery

res

Resolution of the aerial imagery. Units are in denoted in the units column.

units

Units of the resolution aerial imagery

year

Year of the aerial imagery

geometry

Geometry of the aerial imagery

Source

Package development.