Calibrate AEME model parameters using observations

calib_model() runs the model and compares it against observations provided. It can run in parallel by using multiple cores availlable on your computer to run quicker.

Usage

calib_aeme(
  aeme_data,
  path,
  param,
  model,
  mod_ctrls,
  vars_sim = "HYD_temp",
  FUN = NULL,
  ctrl = NULL,
  weights = c(1),
  param_df = NULL
)

Arguments

aeme_data

aeme; data object.

path

filepath; where input files are located relative to `config`.

param

dataframe; of parameters read in from a csv file. Requires the columns c("model", "file", "name", "value", "min", "max", "log")

model

string; for which model. Options are c("dy_cd", "glm_aed" and "gotm_wet")

mod_ctrls

dataframe; of configuration loaded from "mod_ctrls.csv".

vars_sim

vector; of variables names to be used in the calculation of model fit. Currently only supports using one variable.

FUN

function; of the form function(O, P) which will be used in to calculate model fit. If NULL, uses mean absolute error (MAE).

ctrl

list; of controls for calibration function.

VTR Value to be reached. The optimization process will stop if either the maximum number of iterations itermax is reached or the best parameter vector bestmem has found a value fn(bestmem) <= VTR. Default to -Inf.

NP number of population members. Defaults to NA; if the user does not change the value of NP from NA it is reset as 10 * sum(param$model == model). For many problems it is best to set NP to be at least 10 times the length of the parameter vector.

itermax the maximum iteration (population generation) allowed. Default is 200.

reltol relative convergence tolerance. The algorithm stops if it is unable to reduce the value by a factor of reltol * (abs(val) + reltol). Default = 0.07

p: The quantile used to select the parents for the next generation. For example, if p = 0.25, the best 25% of the population will be used as parents for the next generation.

mutate fraction of population to undergo mutation (0-1).

parallel boolean; run calibration in parallel. Default to TRUE

out_file filepath; to csv for calibration output to be written to. Defaults to "results.csv"

na_value value to replace NA values with in observations. Default to

ncore: The number of cores to use for the calibration. This is only used if parallel = TRUE. Default to parallel::detectCores() - 1.

weights

vector; of weights for each variable in vars_sim. Default to c(1).

param_df

dataframe; of parameters read in from a csv file. Requires the columns c("model", "file", "name", "value", "min", "max").

Value

list; ctrl which was supplied with updated arguments if missing.