Run sensitivity analysis on AEME model parameters

calib_model() runs the model and compares it against observations provided. It can run in parallel by using multiple cores availlable on your computer to run quicker.

Usage

sa_aeme(
  aeme_data,
  path = ".",
  param,
  model,
  mod_ctrls,
  FUN_list = NULL,
  ctrl = NULL,
  param_df = NULL
)

Arguments

aeme_data

aeme; data object.

path

filepath; where input files are located relative to `config`.

param

dataframe; of parameters read in from a csv file. Requires the columns c("model", "file", "name", "value", "min", "max", "log")

model

string; for which model. Options are c("dy_cd", "glm_aed" and "gotm_wet")

mod_ctrls

dataframe; of configuration loaded from "mod_ctrls.csv".

FUN_list

list of functions; named according to the variables in the vars_sim. Funtions are of the form function(df) which will be used to calculate model fit. If NULL, uses mean absolute error (MAE).

ctrl

list; of controls for calibration function. The control parameters are as follows:

N: The initial sample size of the base sample matrix.

ncore: The number of cores to use for the calibration. This is only used if parallel = TRUE. Default to parallel::detectCores() - 1.

na_value: value to replace NA value when returned.

parallel: Logical value. If TRUE, the sensitivity analysis will be run in parallel. If FALSE, it will be run in series.

out_file: A character string naming a file for writing. Currently it can write to a .csv file or a database (.db; using the duckDB package) file. If writing to a database, it creates a table named "sa_output".

vars_sim: A named list of output variables for sensitivity analysis. The name is user defined but each list must contain:

var: The variable name to use for the sensitivity analysis.

month: A vector of months to use for the sensitivity analysis.

depth_range: A vector of length 2 with the minimum and maximum depth range to use for the sensitivity analysis.

param_df

dataframe; of parameters read in from a csv file. Requires the columns c("model", "file", "name", "value", "min", "max").

Value

list; ctrl which was supplied with updated arguments if missing.

Examples

if (FALSE) {
  # Run sensitivity analysis
  tmpdir <- tempdir()
  aeme_dir <- system.file("extdata/lake/", package = "AEME")
  # Copy files from package into tempdir
  file.copy(aeme_dir, tmpdir, recursive = TRUE)
  path <- file.path(tmpdir, "lake")
  aeme_data <- AEME::yaml_to_aeme(path = path, "aeme.yaml")
  mod_ctrls <- read.csv(file.path(path, "model_controls.csv"))
  inf_factor = c("dy_cd" = 1, "glm_aed" = 1, "gotm_wet" = 1)
  outf_factor = c("dy_cd" = 1, "glm_aed" = 1, "gotm_wet" = 1)
  model <- c("glm_aed")
  aeme_data <- AEME::build_ensemble(path = path, aeme_data = aeme_data,
                                    model = model, mod_ctrls = mod_ctrls,
                                    inf_factor = inf_factor, ext_elev = 5,
                                    use_bgc = FALSE)

  # Load parameters
  utils::data("aeme_parameters", package = "aemetools")
  param <- aeme_parameters |>
    dplyr::filter(file != "wdr")

  # Function to calculate fitness
  fit <- function(df) {
    mean(df$model)
  }

  # Assign function to variable
  FUN_list <- list(HYD_temp = fit)

  # Set up control parameters for surface and bottom temperature
  ctrl <- list(N = 2^3, ncore = 2L, na_value = 999, parallel = TRUE,
               out_file = "results.db",
               vars_sim = list(
                 surf_temp = list(var = "HYD_temp",
                                  month = c(10:12, 1:3),
                                  depth_range = c(0, 2)
                 ),
                 bot_temp = list(var = "HYD_temp",
                                 month = c(10:12, 1:3),
                                 depth_range = c(10, 13)
                 )
               )
  )

  # Run sensitivity analysis AEME model
  ctrl <- sa_aeme(aeme_data = aeme_data, path = path, param = param,
                  model = model, ctrl = ctrl, mod_ctrls = mod_ctrls,
                  FUN_list = FUN_list)
}